+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/core/ModRecipes.java
@@ -1,10 +1,7 @@
 package com.jaquadro.minecraft.storagedrawers.core;
 
 import com.jaquadro.minecraft.storagedrawers.ModConstants;
-import com.jaquadro.minecraft.storagedrawers.core.recipe.AddUpgradeRecipe;
-import com.jaquadro.minecraft.storagedrawers.core.recipe.KeyringRecipe;
-import com.jaquadro.minecraft.storagedrawers.core.recipe.RemoteGroupUpgradeRecipe;
-import com.jaquadro.minecraft.storagedrawers.core.recipe.UpgradeDetachedDrawerRecipe;
+import com.jaquadro.minecraft.storagedrawers.core.recipe.*;
 import com.texelsaurus.minecraft.chameleon.ChameleonServices;
 import com.texelsaurus.minecraft.chameleon.api.ChameleonInit;
 import com.texelsaurus.minecraft.chameleon.registry.ChameleonRegistry;
@@ -21,6 +18,7 @@ public class ModRecipes
     public static final RegistryEntry<RecipeSerializer<KeyringRecipe>> KEYRING_RECIPE_SERIALIZER = RECIPES.register("keyring", () -> new SimpleCraftingRecipeSerializer<>(KeyringRecipe::new));
     public static final RegistryEntry<RecipeSerializer<RemoteGroupUpgradeRecipe>> REMOTE_GROUP_UPGRADE_SERIALIZER = RECIPES.register("remote_group_upgrade", () -> new SimpleCraftingRecipeSerializer<>(RemoteGroupUpgradeRecipe::new));
     public static final RegistryEntry<RecipeSerializer<UpgradeDetachedDrawerRecipe>> DETACHED_UPGRADE_RECIPE_SERIALIZER = RECIPES.register("add_detached_upgrade", () -> new SimpleCraftingRecipeSerializer<>(UpgradeDetachedDrawerRecipe::new));
+    public static final RegistryEntry<RecipeSerializer<PersonalKeyRecipe>> PERSONAL_KEY_RECIPE_SERIALIZER = RECIPES.register("personal_key_cycle", () -> new SimpleCraftingRecipeSerializer<>(PersonalKeyRecipe::new));
 
     public static void init (ChameleonInit.InitContext context) {
         RECIPES.init(context);