+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/item/ItemPersonalKey.java
@@ -2,6 +2,7 @@ package com.jaquadro.minecraft.storagedrawers.item;
 
 import com.jaquadro.minecraft.storagedrawers.api.storage.IDrawerAttributesModifiable;
 import com.jaquadro.minecraft.storagedrawers.config.ModCommonConfig;
+import com.jaquadro.minecraft.storagedrawers.core.ModSecurity;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.Item;
@@ -27,8 +28,10 @@ public class ItemPersonalKey extends ItemKey
 
     @Override
     public boolean isEnabled () {
-        if (securityProvider != null && securityProvider.equals("cofh"))
-            return false;
+        if (securityProvider != null && !securityProvider.equals("unlock")) {
+            if (ModSecurity.registry.getProvider(securityProvider) == null)
+                return false;
+        }
         return ModCommonConfig.INSTANCE.TOOLS.personalKey.enable.get();
     }
 }