+++ fabric/src/main/java/com/jaquadro/minecraft/storagedrawers/integration/Waila.java
@@ -1,5 +1,6 @@
 package com.jaquadro.minecraft.storagedrawers.integration;
 
+import com.jaquadro.minecraft.storagedrawers.ModConstants;
 import com.jaquadro.minecraft.storagedrawers.StorageDrawers;
 import com.jaquadro.minecraft.storagedrawers.block.BlockDrawers;
 import com.jaquadro.minecraft.storagedrawers.block.tile.BlockEntityDrawers;
@@ -13,7 +14,7 @@ import snownee.jade.api.config.IPluginConfig;
 import snownee.jade.api.ui.IElement;
 import snownee.jade.impl.ui.ItemStackElement;
 
-@WailaPlugin(StorageDrawers.MOD_ID)
+@WailaPlugin(ModConstants.MOD_ID)
 public class Waila implements IWailaPlugin
 {
     @Override
@@ -22,9 +23,9 @@ public class Waila implements IWailaPlugin
             || !ModClientConfig.INSTANCE.INTEGRATION.enableWaila.get())
             return;
 
-        registration.addConfig(StorageDrawers.rl("display.content"), true);
-        registration.addConfig(StorageDrawers.rl("display.stacklimit"), true);
-        registration.addConfig(StorageDrawers.rl("display.status"), true);
+        registration.addConfig(ModConstants.loc("display.content"), true);
+        registration.addConfig(ModConstants.loc("display.stacklimit"), true);
+        registration.addConfig(ModConstants.loc("display.status"), true);
 
         WailaDrawer provider = new WailaDrawer();
         registration.registerBlockComponent(provider, BlockDrawers.class);
@@ -43,16 +44,16 @@ public class Waila implements IWailaPlugin
             BlockEntityDrawers blockEntityDrawers = (BlockEntityDrawers) accessor.getBlockEntity();
 
             DrawerOverlay overlay = new DrawerOverlay();
-            overlay.showContent = false; // config.get(StorageDrawers.rl("display.content"));
-            overlay.showStackLimit = config.get(StorageDrawers.rl("display.stacklimit"));
-            overlay.showStatus = config.get(StorageDrawers.rl("display.status"));
+            overlay.showContent = config.get(ModConstants.loc("display.content"));
+            overlay.showStackLimit = config.get(ModConstants.loc("display.stacklimit"));
+            overlay.showStatus = config.get(ModConstants.loc("display.status"));
 
             currenttip.addAll(overlay.getOverlay(blockEntityDrawers));
         }
 
         @Override
         public ResourceLocation getUid () {
-            return new ResourceLocation(StorageDrawers.MOD_ID, "main");
+            return new ResourceLocation(ModConstants.MOD_ID, "main");
         }
     }
 }