+++ forge/src/main/java/com/jaquadro/minecraft/storagedrawers/StorageDrawers.java
@@ -3,6 +3,7 @@ package com.jaquadro.minecraft.storagedrawers;
 import com.jaquadro.minecraft.storagedrawers.capabilities.PlatformCapabilities;
 import com.jaquadro.minecraft.storagedrawers.config.*;
 import com.jaquadro.minecraft.storagedrawers.core.*;
+import com.jaquadro.minecraft.storagedrawers.integration.LocalIntegrationRegistry;
 import com.jaquadro.minecraft.storagedrawers.network.PlayerBoolConfigMessage;
 import com.texelsaurus.minecraft.chameleon.ChameleonServices;
 import com.texelsaurus.minecraft.chameleon.registry.ForgeRegistryContext;
@@ -80,6 +81,10 @@ public class StorageDrawers
         MaterialBlacklist.INSTANCE.initialize();
         ConversionRegistry.INSTANCE.initialize();
         PlatformCapabilities.initHandlers();
+
+        LocalIntegrationRegistry.initialize();
+        LocalIntegrationRegistry.instance().init();
+        LocalIntegrationRegistry.instance().postInit();
 
         //oreDictRegistry = new OreDictRegistry();
         //renderRegistry = new RenderRegistry();